Moved models.go in internal\app\models
